refactor: Next Warning Card component
